pyproject.toml
[project]
name = "k8s-mcp-server"
version = "1.0.0"
description = "Kubernetes MCP Server - 提供Kubernetes集群管理的MCP服务"
authors = [
    {name = "K8s MCP Server Team", email = "team@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastmcp>=2.0.0",
    "kubernetes>=25.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.20.0",
    "PyJWT>=2.8.0",
]

[project.scripts]
mcp-admin = "mcp_admin_cli:main"

[tool.setuptools]
py-modules = ["mcp_admin_cli"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/your-org/k8s-mcp-server"
Repository = "https://github.com/your-org/k8s-mcp-server"
Issues = "https://github.com/your-org/k8s-mcp-server/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*", "services*", "utils*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 